+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Ensure zero value duration is correctly formatted [#683](https://github.com/markuspoerschke/iCal/pull/683)
+
 ## [2.16.0] - 2026-01-07
 
 ### Added